+++ packages/ui-components/src/index.ts
@@ -1,5 +1,5 @@
-export { ActionButton } from "./uiComponents/ActionButton";
-export { ActionButtonGroup } from "./uiComponents/ActionButtonGroup";
-export { Button } from "./uiComponents/Button";
-export { ToggleButton } from "./uiComponents/ToggleButton";
-export { ToggleButtonGroup } from "./uiComponents/ToggleButtonGroup";
+export { ActionButton } from "./components/ActionButton";
+export { ActionButtonGroup } from "./components/ActionButtonGroup";
+export { Button } from "./components/Button";
+export { ToggleButton } from "./components/ToggleButton";
+export { ToggleButtonGroup } from "./components/ToggleButtonGroup";
